% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestcluster.R
\name{bestcluster}
\alias{bestcluster}
\title{Best Clustering by Least Squared Distance to Probability Matrix}
\usage{
bestcluster(bpr)
}
\arguments{
\item{bpr}{object of class bpr; output from function profileReg}
}
\value{
vector denoting the best clustering of the data using least squares distance to a similarity matrix as described in Dahl (2006)
}
\description{
Best Clustering by Least Squared Distance to Probability Matrix
}
\references{
Dahl, D. B. (2006). Model-Based Clustering for Expression Data via a Dirichlet Process Mixture Model. Bayesian Inference for Gene Expression and Proteomics, 201–218. https://doi.org/10.1017/CBO9780511584589.011
}
