% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bkmr_wrapper.R
\name{bkmr_wrapper}
\alias{bkmr_wrapper}
\title{Wrapper function to fit bkmr model in package bkmr.}
\usage{
bkmr_wrapper(niter, nburn, Y, X, W, varsel = FALSE, groups = NULL)
}
\arguments{
\item{niter}{number of iterations of sampler}

\item{nburn}{number of burn-in iterations}

\item{Y}{n x 1 vector of response data}

\item{X}{n x p matrix of exposure data}

\item{W}{n x q matrix of covariate data}

\item{varsel}{logical; if TRUE variable selection is implemented, default is FALSE}

\item{groups}{optional vector of length p indicating group membership to conduct hierarchical variable selection}
}
\value{
a list with components 
\itemize{
    \item fit: the original model fit 
    \item risk: the posterior mean risk for each subject
    \item pips: component-wise (if groups = NULL) or conditional (if groups is specified) posterior inclusion probabilities for exposures
    \item group.pips: group posterior inclusion probabilities (only if groups is specified))
    \item risk.summary: mean, SD, and 0.95 CI for risk for each subject 
    \item preds: predicted values for each subject based on exposures and covariates
    \item risk.distn: posterior distribution of risk for each subject
}
}
\description{
Wrapper function to fit bkmr model in package bkmr.
}
