% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brier.R
\name{brier}
\alias{brier}
\title{Calculate brier scores for Bayesian methods}
\usage{
brier(X, active, pips)
}
\arguments{
\item{X}{exposure data}

\item{active}{true active exposures}

\item{pips}{posterior inclusion probabilities from model fit}
}
\value{
brier score
}
\description{
Calculate brier scores for Bayesian methods
}
