% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModels.R
\name{fitModels}
\alias{fitModels}
\title{Fit all 5 Bayesian methods and summarize output}
\usage{
fitModels(names = NULL, simnum, niter, nburn, X, W, data, seed = NULL)
}
\arguments{
\item{names}{optional name for the simulation (i.e. "linear")}

\item{simnum}{simulation number}

\item{niter}{number of total iterations for each method}

\item{nburn}{number of burn-in iterations for each method}

\item{X}{exposure data}

\item{W}{covariate data}

\item{data}{list returned from function simLinearResponse, simNonlinearResponse, or simProfilesResponse}

\item{seed}{seed}
}
\value{
data frame summarizing each method
}
\description{
Fit all 5 Bayesian methods and summarize output
}
