% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logsum.R
\name{logsum}
\alias{logsum}
\title{Log of sum of exponentiated log-likelihood}
\usage{
logsum(x)
}
\arguments{
\item{x}{vector of log-likelihoods}
}
\value{
return log of the sum of the exponentiated log-likelihoods
}
\description{
Log of sum of exponentiated log-likelihood
}
