% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelaves.R
\name{modelaves}
\alias{modelaves}
\title{Model Average Cluster-Specific Intercepts for Profile Regression}
\usage{
modelaves(bpr)
}
\arguments{
\item{bpr}{object of class bpr}
}
\value{
summary of model averaged posterior estimates for cluster-specific parameters
}
\description{
computes model averaged cluster-specific intercepts (theta) from supervised or unsupervised Bayesian profile regression model
}
