% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npb.R
\name{npb}
\alias{npb}
\title{Function to fit nonparametric Bayes shrinkage}
\usage{
npb(niter, nburn, X, Y, W, scaleY = FALSE, priors, interact = FALSE,
  intercept = TRUE)
}
\arguments{
\item{niter}{number of iterations}

\item{nburn}{number of burn-in iterations}

\item{X}{n by p matrix of predictor data}

\item{Y}{n by 1 vector of continuous response data}

\item{W}{n by q matrix of covariate data}

\item{scaleY}{logical; if TRUE response will be centered and scaled before model fit, default is FALSE}

\item{priors}{list of prior hyperparameters, see package documentation for details}

\item{interact}{logical; if TRUE (default) include all pairwise interactions of predictors, if FALSE include main effects only}

\item{intercept}{logical; indicates if an overall intercept should be estimated with covariates}
}
\value{
an object of class "npb", which has the associated methods:
\itemize{
\item \code{\link{print}} (i.e. \code{\link{print.npb}} )
\item \code{\link{summary}} (i.e. \code{\link{summary.npb}} )
\item \code{\link{predict}} (i.e. \code{\link{predict.npb}} )
}

a list with components
\itemize{
   \item X: predictor data matrix
   \item Y: response data vector
   \item Z: matrix of pairwise multiplicative interactions of predictor data (only if interact = TRUE)
   \item W: covariate data
   \item alpha: DP parameter for main effects estimates 
   \item alpha.2: DP parameter for interactions estimates 
   \item beta: unscaled main effect regression coefficient estimates 
   \item zeta: unscaled interaction regression coefficient estimates 
   \item mu: mu estimates 
   \item mu.2: mu2 estimates 
   \item phi2inv: phi2inv estimates
   \item phi2inv.2: phi2inv.2 estimates 
   \item gamma: unscaled covariate regression coefficient estimates 
   \item sig2inv: sig2inv estimates 
   \item pip.beta: inclusion indicators for main effects
   \item pip.zeta: inclusion indicators for interactions
}
}
\description{
Fits nonparametric Bayes shrinkage model for continuous response data using Markov chain Monte Carlo (MCMC) methods
}
