% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npb_int.R
\name{npb_int}
\alias{npb_int}
\title{Fit nonparametric Bayes shrinkage model with interactions}
\usage{
npb_int(niter, nburn, X, Y, W, scaleY = FALSE, priors)
}
\arguments{
\item{niter}{number of iterations}

\item{nburn}{number of burn-in iterations}

\item{X}{n by p matrix of predictor data}

\item{Y}{n by 1 vector of response data}

\item{W}{n by q matrix of covariate data}

\item{scaleY}{logical; if TRUE response will be centered and scaled before model fit}

\item{priors}{list of prior hyperparameters}
}
\value{
list of model estimates, see npb documentation
}
\description{
Fit nonparametric Bayes shrinkage model with interactions
}
