% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.npb.R
\name{predict.npb}
\alias{predict.npb}
\title{Predict method for class "npb"}
\usage{
\method{predict}{npb}(object, ...)
}
\arguments{
\item{object}{object of class "npb"}

\item{...}{ignored}
}
\value{
list with components
\itemize{
     \item fitted.vals: posterior mean fitted values for each subject
     \item fitted.distn: posterior distribution of fitted values for each subject  
}
}
\description{
Predict method for class "npb"
}
