% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/premium_wrapper.R
\name{premium_wrapper}
\alias{premium_wrapper}
\title{Wrapper function to fit profile regression model from package PReMiuM}
\usage{
premium_wrapper(
  niter,
  nburn,
  Y,
  X,
  W,
  scaleY = FALSE,
  varSelectType = "None",
  simnum = NULL,
  priors,
  seed = NULL
)
}
\arguments{
\item{niter}{number of iterations through sampler}

\item{nburn}{number of burn-in iterations}

\item{Y}{response data}

\item{X}{exposure data}

\item{W}{covariate data}

\item{scaleY}{logical; if TRUE response is centered and scaled before model fit}

\item{varSelectType}{type of variable selection to be used: "None" or "BinaryCluster", default is "None"}

\item{simnum}{number of output file (.txt files will be saved in the file path paste0("Premium_output/output",simnum)}

\item{priors}{prior hyperparameters}

\item{seed}{random number seed}
}
\value{
list with components
\itemize{
    \item fit: model fit, an object of type "runInfoObj". See profRegr documentation in R package PReMiuM
    \item risk: mean risk for each subject
    \item rho: posterior distribution of probability of inclusion for each exposure 
    \item risk.summary: mean, SD, and 0.95 CI of cluster-intercept (risk) for each subject
    \item exposure.response: cluster means, SD, 0.95 CI, and size, plus mean/SD of repsonse for subjects in each cluster  
    \item fitted vals: fitted values for each subject
    \item clusters: vector of most optimal clustering 
    \item cluster.summary: array of empirical mean and SD of exposures for individuals assigned to each cluster
    \item risk distn: distribution of model-averaged cluster-intercepts for each cluster in the best clustering
    \item groupList: list of which subjects are in which group in the most optimal clustering 
    \item riskProfileObj: object of type "riskProfileObj". See PReMiuM documentation for function "calcAvgRiskAndProfile"
  
}
}
\description{
Wrapper function to fit profile regression model from package PReMiuM
}
