% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.bpr.R
\name{print.bpr}
\alias{print.bpr}
\title{Print method for class "bpr"}
\usage{
\method{print}{bpr}(x, ...)
}
\arguments{
\item{x}{object of class bpr}

\item{...}{ignored}
}
\value{
call to summary.bpr
}
\description{
Print method for class "bpr"
}
