% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.npb.R
\name{print.npb}
\alias{print.npb}
\title{Print method for class "npb"}
\usage{
\method{print}{npb}(x, ...)
}
\arguments{
\item{x}{object of class "npb"}

\item{...}{ignored}
}
\value{
list of output from summary.npb
}
\description{
Print method for class "npb"
}
