% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.bpr.R
\name{print.summary.bpr}
\alias{print.summary.bpr}
\title{Print method for class  "summary.bpr"}
\usage{
\method{print}{summary.bpr}(x, ...)
}
\arguments{
\item{x}{object of class summary.bpr}

\item{...}{ignored}
}
\value{
call to summary.bpr
}
\description{
Print method for class  "summary.bpr"
}
