% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.npb.R
\name{print.summary.npb}
\alias{print.summary.npb}
\title{Print method for class "summary.npb"}
\usage{
\method{print}{summary.npb}(x, ...)
}
\arguments{
\item{x}{object of class summary.npb}

\item{...}{ignored}
}
\value{
call to summary.npb
}
\description{
Print method for class "summary.npb"
}
