% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileReg.R
\name{profileReg}
\alias{profileReg}
\title{Fit Bayesian profile regression}
\usage{
profileReg(niter, nburn, X, Y, W, C = 20, scaleY = FALSE,
  DPgamma = TRUE, varsel = FALSE, priors, sup = TRUE)
}
\arguments{
\item{niter}{number of total iterations}

\item{nburn}{number of burn-in iterations}

\item{X}{n by p matrix of predictor data}

\item{Y}{n by 1 vector of continuous response data}

\item{W}{n by q matrix of covariate data}

\item{C}{maximum number of clusters allowed}

\item{scaleY}{logical; if TRUE response will be centered and scaled before model fit}

\item{DPgamma}{logical; if TRUE (default) alpha has a gamma prior, else alpha has Unif(.03, 10) prior}

\item{varsel}{logical; if TRUE binary cluster variable selection is implemented (see Chung and Dunson 2009), default is FALSE}

\item{priors}{list of prior hyperparameters, see package documentation for details}

\item{sup}{logical; if TRUE (default) fits supervised model, else fits unsupervised model}
}
\value{
list with components
\itemize{
   \item X: predictor data matrix
   \item W: covariate data matrix
   \item Y: response data vector
   \item C: maximum number of allowable clusters
   \item alpha: DP parameter estimates
   \item mu: array of cluster means estimates
   \item psi: cluster weights estimates
   \item Z: cluster indicators at each iteraction
   \item delta: regression coefficient estimates for theta (risk) and gamma (fixed effects)
   \item sig2inv: error precision estimates
   \item kap2inv: cluster intercept (risk) precision estimates
   \item phi2inv: fixed effect precision estimates
   \item rho: rho estimates
}
}
\description{
Fits the Bayesian profile regression model using MCMC methods
}
