% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcat.R
\name{rcat}
\alias{rcat}
\title{Random categorical generator}
\usage{
rcat(probs)
}
\arguments{
\item{probs}{n by K matrix of probabilities n subjects to be assigned to K categories}
}
\value{
return vector of length nrow(probs) of categorical indicators
}
\description{
Random categorical generator
}
