% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simexpodat.R
\name{simexpodat}
\alias{simexpodat}
\title{Simulate Data for MixModelPack}
\usage{
simexpodat(n, Xdat)
}
\arguments{
\item{n}{sample size}

\item{Xdat}{Xdat exposure data matrix loaded from package mmpack}
}
\value{
list with components
\itemize{
   \item Y: response data
   \item X: exposure data
   \item W: covariate data
   \item h: exposure-response function
   \item active: active main effects
   \item active.ints: active interactions 
}
}
\description{
Function to simulate exposure, covariate, and response data for up to 1000 observations
}
