% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumERbayes.R
\name{sumERbayes}
\alias{sumERbayes}
\title{Summarize evaluation of estimated exposure-response function for Bayesian methods}
\usage{
sumERbayes(fit, h)
}
\arguments{
\item{fit}{fit or summary from one of the 5 Bayesian methods}

\item{h}{true exposure-response function}
}
\value{
summary of estimated exposure-response function
}
\description{
Summarize evaluation of estimated exposure-response function for Bayesian methods
}
