% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumERlinearmodels.R
\name{sumERlinearmodels}
\alias{sumERlinearmodels}
\title{Summarize evaluation of estimated exposure-response function for linear models}
\usage{
sumERlinearmodels(fit, h, interaction, X, W)
}
\arguments{
\item{fit}{object of type lm}

\item{h}{true exposure-response function}

\item{interaction}{logical; was LM fit with interactions?}

\item{X}{exposure data}

\item{W}{covariate data}
}
\value{
summary of estimated exposure-response function
}
\description{
Summarize evaluation of estimated exposure-response function for linear models
}
