% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeSimulation.R
\name{summarizeSimulation}
\alias{summarizeSimulation}
\title{Summarize results from evaluating methods through simulation with function fitModels}
\usage{
summarizeSimulation(df)
}
\arguments{
\item{df}{results data frame as output from function fitModels}
}
\value{
data frame of simulation results averaged over number of data sets
}
\description{
Summarize results from evaluating methods through simulation with function fitModels
}
