% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bpr.R
\name{summary.bpr}
\alias{summary.bpr}
\title{Summary method for class "bpr"}
\usage{
\method{summary}{bpr}(object, ...)
}
\arguments{
\item{object}{object of class "bpr"}

\item{...}{ignored}
}
\value{
list of summary data with components
\itemize{
     \item exposure.response: cluster means, SD, 0.95 CI, and size, plus mean/SD of repsonse for subjects in each cluster 
     \item risk: mean risk for each subject
     \item cluster.summary: array of empirical mean and SD of exposures for individuals assigned to each cluster
     \item risk.summary: mean, SD, and 0.95 CI of cluster-intercept (risk) for each subject
     \item clusters: vector of most optimal clustering  
     \item risk.distn: distribution of model-averaged cluster-intercepts for each cluster in the best clustering
     \item groupList: list of which subjects belong to which cluster in the best clustering 
     \item rho: posterior mean probability of inclusion for each exposure 
}
}
\description{
Summary method for class "bpr"
}
