% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.npb.R
\name{summary.npb}
\alias{summary.npb}
\title{Summary method for class "npb"}
\usage{
\method{summary}{npb}(object, ...)
}
\arguments{
\item{object}{object of class "npb"}

\item{...}{ignored}
}
\value{
list of summary data with components
\itemize{
    \item main.effects: mean, sd, and 0.95 CI for main effects
    \item interactions: mean, sd, and 0.95 CI for interactions
    \item selected.main: mean, sd, and 0.95 CI for main effects with PIPs > 0.5
    \item selected.int: mean, sd, and 0.95 CI for interactions with PIPs > 0.5
    \item risk.summary: mean, sd and 0.95 CI for exposure-response function "risk"
    \item risk: mean risk for each individual
    \item risk.distn: matrix, columns are the posterior distribution of risk for each individual 
}
}
\description{
Summary method for class "npb"
}
