% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_S.R
\name{update_S}
\alias{update_S}
\title{Update allocation variables in NPB}
\usage{
update_S(S, p, alpha.pi, beta.pi, beta, theta, zeta, Y, W, gamma, X, Z,
  sig2inv, phi2inv, mu, alpha, iter)
}
\arguments{
\item{S}{allocation variable to update}

\item{p}{number of parameters}

\item{alpha.pi}{shape1 parameter for beta prior on pi0}

\item{beta.pi}{shape2 parameter for beta prior on pi0}

\item{beta}{regression coefficients corresponding to S}

\item{theta}{clusters corresponding to S}

\item{zeta}{other regression coefficients}

\item{Y}{response}

\item{W}{covariates}

\item{gamma}{covariate regression coefficients}

\item{X}{data for regression coefficients being updated}

\item{Z}{data for regression coefficients NOT being updated}

\item{sig2inv}{error precision}

\item{phi2inv}{precision for theta}

\item{mu}{mean for theta}

\item{alpha}{DP parameter for theta}

\item{iter}{MCMC iteration for debugging purposes}
}
\value{
vector of updated allocation variables that assign each regression coefficient (beta) to a specific cluster (theta)
}
\description{
Function to update the allocations of regression coefficients to clusters in NPB model
}
