% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xdat.R
\docType{data}
\name{Xdat}
\alias{Xdat}
\title{Exposure data}
\format{
A matrix with 1000 rows and 7 columns
\describe{
  \item{C}{carbamates, kg/(km^2)}
  \item{MeBr}{methyl, bromide kg/(km^2)}
  \item{OP}{organophosphates, kg/(km^2)}
  \item{NO2}{nitrogen dioxide, ppb}
  \item{O3}{ozone, ppm}
  \item{PM10}{coarse particulate matter, micograms/m^3}
  \item{PM2.5}{fine particulate matter, micograms/m^3}
}
}
\usage{
data(Xdat)
}
\description{
A dataset containing 1000 simulated values for C, MeBr, OP, NO2, O3, PM10, and PM2.5
}
\examples{
data(Xdat)



}
\keyword{datasets}
